requirements.txt
flake8
pytest
openpyxl
requests
python-dotenv
